Update set_period for tone in mockingboard

+++ emulator/src/mockingboard.rs
@@ -148,7 +148,7 @@ impl Tone {
         if self.period == 0 {
             self.period = 1;
         }
-        self.count %= self.period;
+        self.count %= self.period as usize;
     }
 
     fn set_volume(&mut self, val: u8) {
